fix: Bin shape groups
Without binning whole shape groups, the style information of the last
shape in the group differs per tile, causing inconsistencies in the
styling.

This fixes the issue by binning groups instead of shapes. This may have
a big performance hit in some cases, which could be addressed in the
future by not binning the entire group, but only the overlapping shapes
as well as the last shape (which provides style information)